requirements.txt
# Core dependencies
python-dotenv==1.1.1
pyodbc==5.2.0

# Google Gemini AI
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10

# JWT Authentication
PyJWT==2.9.0

# Web Framework
fastapi==0.115.14
uvicorn==0.35.0

# Optional: Testing
# pytest==8.3.4
# pytest-cov==6.0.0
